refactor: separate retained developer tooling tests
