+++ package.json
@@ -17,8 +17,8 @@
     "test:run": "vitest run --coverage && node scripts/check-coverage.mjs --threshold 90 --branches 85",
     "lint": "oxlint src/",
     "e2e:typecheck": "tsc --project tsconfig.e2e.json --noEmit",
-    "e2e:lint": "oxlint test/e2e/ test/native/ test/support/ scripts/*.mjs",
-    "e2e:format:check": "prettier --check test/e2e/ test/native/ test/support/ test/fixtures/ scripts/*.mjs tsconfig.e2e.json",
+    "e2e:lint": "oxlint test/e2e/ test/native/ test/support/ test/tooling/ scripts/*.mjs",
+    "e2e:format:check": "prettier --check test/e2e/ test/native/ test/support/ test/tooling/ test/fixtures/ scripts/*.mjs tsconfig.e2e.json",
     "docs:format:check": "prettier --check --ignore-path .gitignore AGENTS.md ARCHITECTURE.md CONVENTIONS.md DEVELOPMENT.md RUST-REWRITE.md PERFORMANCE-BASELINE.md \".agents/skills/**/SKILL.md\" .github/pull_request_template.md",
     "lint:fix": "oxlint src/ --fix",
     "format": "prettier --write src/",
