+++ test/support/source-imports.ts
@@ -1,0 +1,40 @@
+import ts from 'typescript';
+
+// Test-only literal module edge extraction; shared by product and tooling guards.
+export function imports(text: string): string[] {
+  const source = ts.createSourceFile('module.ts', text, ts.ScriptTarget.Latest, true);
+  const values: string[] = [];
+  const collect = (node: ts.Node): void => {
+    if (
+      (ts.isImportDeclaration(node) || ts.isExportDeclaration(node)) &&
+      node.moduleSpecifier &&
+      ts.isStringLiteral(node.moduleSpecifier)
+    ) {
+      values.push(node.moduleSpecifier.text);
+    } else if (
+      ts.isCallExpression(node) &&
+      (node.expression.kind === ts.SyntaxKind.ImportKeyword ||
+        (ts.isIdentifier(node.expression) && node.expression.text === 'require')) &&
+      node.arguments.length === 1 &&
+      ts.isStringLiteral(node.arguments[0])
+    ) {
+      values.push(node.arguments[0].text);
+    } else if (
+      ts.isImportTypeNode(node) &&
+      ts.isLiteralTypeNode(node.argument) &&
+      ts.isStringLiteral(node.argument.literal)
+    ) {
+      values.push(node.argument.literal.text);
+    } else if (
+      ts.isImportEqualsDeclaration(node) &&
+      ts.isExternalModuleReference(node.moduleReference) &&
+      node.moduleReference.expression &&
+      ts.isStringLiteral(node.moduleReference.expression)
+    ) {
+      values.push(node.moduleReference.expression.text);
+    }
+    ts.forEachChild(node, collect);
+  };
+  collect(source);
+  return values;
+}
