+++ test/tooling/native-artifact-policy.test.ts
@@ -7,9 +7,9 @@ import { fileURLToPath, pathToFileURL } from 'node:url';
 import { createGzip } from 'node:zlib';
 import { describe, expect, it } from 'vitest';
 import * as tar from 'tar';
-import { runCli, withSandbox, type Sandbox } from '../test/support/cli-process.js';
+import { runCli, withSandbox, type Sandbox } from '../support/cli-process.js';
 
-const repositoryRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
+const repositoryRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..', '..');
 const { selectNativeArtifact, withNativeArtifact } = (await import(
   pathToFileURL(path.join(repositoryRoot, 'scripts', 'native-artifact-policy.mjs')).href
 )) as unknown as {
