+++ test/tooling/native-cargo.test.ts
@@ -2,9 +2,9 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { describe, expect, it } from 'vitest';
-import { runCli, withSandbox, type Sandbox } from '../test/support/cli-process.js';
+import { runCli, withSandbox, type Sandbox } from '../support/cli-process.js';
 
-const wrapper = fileURLToPath(new URL('../scripts/native-cargo.sh', import.meta.url));
+const wrapper = fileURLToPath(new URL('../../scripts/native-cargo.sh', import.meta.url));
 
 interface FakeCargo {
   readonly argsFile: string;
