+++ vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['src/**/*.test.ts'],
+    include: ['src/**/*.test.ts', 'test/tooling/**/*.test.ts'],
     exclude: ['test/e2e/**', 'node_modules', 'dist'],
     testTimeout: 1000,
     coverage: {
